+++ tests/test_dq.py
@@ -1,29 +1,6 @@
-""" JWST Data Quality Flags
+from stdatamodels import dqflags
 
-The definitions are documented in the JWST RTD:
-
-https://jwst-pipeline.readthedocs.io/en/latest/jwst/references_general/references_general.html#data-quality-flags
-
-
-Implementation
--------------
-
-The flags are implemented as "bit flags": Each flag is assigned a bit position
-in a byte, or multi-byte word, of memory. If that bit is set, the flag assigned
-to that bit is interpreted as being set or active.
-
-The data structure that stores bit flags is just the standard Python `int`,
-which provides 32 bits. Bits of an integer are most easily referred to using
-the formula `2**bit_number` where `bit_number` is the 0-index bit of interest.
-"""
-
-# These imports are here for backwards compatibility
-from astropy.nddata.bitmask import interpret_bit_flags as ap_interpret_bit_flags
-from stdatamodels.dqflags import interpret_bit_flags, dqflags_to_mnemonics
-from stdatamodels.basic_utils import multiple_replace
-
-# Pixel-specific flags
-pixel = {'GOOD':             0,      # No bits set, all is good
+PIXEL = {'GOOD':             0,      # No bits set, all is good
          'DO_NOT_USE':       2**0,   # Bad pixel. Do not use.
          'SATURATED':        2**1,   # Pixel saturated during exposure
          'JUMP_DET':         2**2,   # Jump detected during exposure
@@ -31,7 +8,7 @@ pixel = {'GOOD':             0,      # No bits set, all is good
          'OUTLIER':          2**4,   # Flagged by outlier detection (was RESERVED_1)
          'PERSISTENCE':      2**5,   # High persistence (was RESERVED_2)
          'AD_FLOOR':         2**6,   # Below A/D floor (0 DN, was RESERVED_3)
-         'UNDERSAMP':        2**7,   # Undersampling correction (was RESERVED_4)
+         'RESERVED_4':       2**7,   #
          'UNRELIABLE_ERROR': 2**8,   # Uncertainty exceeds quoted error
          'NON_SCIENCE':      2**9,   # Pixel not on science portion of detector
          'DEAD':             2**10,  # Dead pixel
@@ -59,16 +36,7 @@ pixel = {'GOOD':             0,      # No bits set, all is good
          }
 
 
-# Group-specific flags. Once groups are combined, these flags
-# are equivalent to the pixel-specific flags.
-group = {'GOOD':       pixel['GOOD'],
-         'DO_NOT_USE': pixel['DO_NOT_USE'],
-         'SATURATED':  pixel['SATURATED'],
-         'JUMP_DET':   pixel['JUMP_DET'],
-         'DROPOUT':    pixel['DROPOUT'],
-         'AD_FLOOR':   pixel['AD_FLOOR'],
-         'UNDERSAMP':  pixel['UNDERSAMP'],
-         }
-
-__all__ = ["ap_interpret_bit_flags", "interpret_bit_flags", "dqflags_to_mnemonics",
-           "multiple_replace", "pixel", "group"]
+def test_dqflags():
+    assert dqflags.dqflags_to_mnemonics(1, PIXEL) == {'DO_NOT_USE'}
+    assert dqflags.dqflags_to_mnemonics(7, PIXEL) == {'JUMP_DET', 'DO_NOT_USE', 'SATURATED'}
+    assert dqflags.interpret_bit_flags('DO_NOT_USE + WARM', mnemonic_map=PIXEL) == 4097