+++ CHANGES.rst
@@ -1,6 +1,7 @@
 1.3.0 (unreleased)
 ==================
 
+
 Other
 -----
 
@@ -8,6 +9,8 @@ Other
   so that the ``stcal`` dependency can be dropped. [#134]
 
 - increase ``requires-python`` to ``3.8`` [#144]
+
+- Add R_MRSXAR as the keyword for the jwst straylight mrsxartcorr reference filename in core schema in stdatamodels.jwst.datamodels [#145]
 
 =======
 Bug Fixes